:shirt: Refactor project structure
