+++ src/SceneGate.UI/Main/ViewKind.cs
@@ -17,7 +17,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace SceneGate.UI.ViewModels
+namespace SceneGate.UI.Main
 {
     public enum ViewKind
     {
